Moved SPA Info in Preparation for Future Changes
### Dev Notes
I'm about to move some logic out of `AbilitiesTab.java` connected to the categorization of SPAs. This will allow that logic to be used anywhere, not just campaign options. However, to do that I needed to move and rename `CampaignOptionsAbilityInfo.java`. I was advised to make a PR that just did the move and renaming, and then to follow up with the logic changes.
