+++ MekHQ/src/mekhq/gui/campaignOptions/contents/AbilitiesTab.java
@@ -24,23 +24,27 @@
  *
  * Catalyst Game Labs and the Catalyst Game Labs logo are trademarks of
  * InMediaRes Productions, LLC.
+ *
+ * MechWarrior Copyright Microsoft Corporation. MekHQ was created under
+ * Microsoft's "Game Content Usage Rules"
+ * <https://www.xbox.com/en-US/developers/rules> and it is not endorsed by or
+ * affiliated with Microsoft.
  */
 package mekhq.gui.campaignOptions.contents;
 
 import static mekhq.campaign.personnel.skills.SkillType.EXP_ELITE;
-import static mekhq.campaign.personnel.skills.SkillType.EXP_NONE;
 import static mekhq.campaign.personnel.skills.SkillType.S_GUN_BA;
 import static mekhq.campaign.personnel.skills.SkillType.S_GUN_PROTO;
 import static mekhq.campaign.personnel.skills.SkillType.getExperienceLevelName;
 import static mekhq.campaign.personnel.skills.enums.SkillSubType.COMBAT_GUNNERY;
 import static mekhq.campaign.personnel.skills.enums.SkillSubType.COMBAT_PILOTING;
-import static mekhq.gui.campaignOptions.CampaignOptionsAbilityInfo.AbilityCategory.CHARACTER_CREATION_ONLY;
-import static mekhq.gui.campaignOptions.CampaignOptionsAbilityInfo.AbilityCategory.CHARACTER_FLAW;
-import static mekhq.gui.campaignOptions.CampaignOptionsAbilityInfo.AbilityCategory.COMBAT_ABILITY;
-import static mekhq.gui.campaignOptions.CampaignOptionsAbilityInfo.AbilityCategory.MANEUVERING_ABILITY;
-import static mekhq.gui.campaignOptions.CampaignOptionsAbilityInfo.AbilityCategory.UTILITY_ABILITY;
 import static mekhq.gui.campaignOptions.CampaignOptionsUtilities.createParentPanel;
 import static mekhq.gui.campaignOptions.CampaignOptionsUtilities.getImageDirectory;
+import static mekhq.utilities.spaUtilities.enums.AbilityInfo.AbilityCategory.CHARACTER_CREATION_ONLY;
+import static mekhq.utilities.spaUtilities.enums.AbilityInfo.AbilityCategory.CHARACTER_FLAW;
+import static mekhq.utilities.spaUtilities.enums.AbilityInfo.AbilityCategory.COMBAT_ABILITY;
+import static mekhq.utilities.spaUtilities.enums.AbilityInfo.AbilityCategory.MANEUVERING_ABILITY;
+import static mekhq.utilities.spaUtilities.enums.AbilityInfo.AbilityCategory.UTILITY_ABILITY;
 
 import java.awt.BorderLayout;
 import java.awt.Dimension;
@@ -53,7 +57,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.ResourceBundle;
-import java.util.regex.Pattern;
 import javax.swing.BorderFactory;
 import javax.swing.JButton;
 import javax.swing.JCheckBox;
@@ -65,22 +68,18 @@ import megamek.client.ui.swing.util.UIUtil;
 import megamek.common.annotations.Nullable;
 import megamek.common.options.IOption;
 import megamek.common.options.IOptionGroup;
-import megamek.logging.MMLogger;
 import mekhq.CampaignPreset;
-import mekhq.MekHQ;
-import mekhq.Utilities;
 import mekhq.campaign.personnel.PersonnelOptions;
 import mekhq.campaign.personnel.SkillPerquisite;
 import mekhq.campaign.personnel.SpecialAbility;
 import mekhq.campaign.personnel.skills.SkillType;
-import mekhq.campaign.personnel.skills.enums.SkillSubType;
-import mekhq.gui.campaignOptions.CampaignOptionsAbilityInfo;
-import mekhq.gui.campaignOptions.CampaignOptionsAbilityInfo.AbilityCategory;
 import mekhq.gui.campaignOptions.components.CampaignOptionsButton;
 import mekhq.gui.campaignOptions.components.CampaignOptionsGridBagConstraints;
 import mekhq.gui.campaignOptions.components.CampaignOptionsHeaderPanel;
 import mekhq.gui.campaignOptions.components.CampaignOptionsStandardPanel;
 import mekhq.gui.dialog.EditSpecialAbilityDialog;
+import mekhq.utilities.spaUtilities.enums.AbilityInfo;
+import mekhq.utilities.spaUtilities.enums.AbilityInfo.AbilityCategory;
 
 /**
  * The {@code AbilitiesTab} class represents a GUI tab for configuring and managing special abilities in a campaign.
@@ -94,7 +93,7 @@ public class AbilitiesTab {
     private static final ResourceBundle resources = ResourceBundle.getBundle(RESOURCE_PACKAGE);
 
     private ArrayList<String> level3Abilities;
-    private Map<String, CampaignOptionsAbilityInfo> allAbilityInfo;
+    private Map<String, AbilityInfo> allAbilityInfo;
     private JPanel combatTab;
     private JPanel maneuveringTab;
     private JPanel utilityTab;
@@ -219,8 +218,7 @@ public class AbilitiesTab {
             }
 
             // Mark the ability as active
-            allAbilityInfo.put(abilityName,
-                  new CampaignOptionsAbilityInfo(abilityName, clonedAbility, isEnabled, category));
+            allAbilityInfo.put(abilityName, new AbilityInfo(abilityName, clonedAbility, isEnabled, category));
         }
     }
 
@@ -308,7 +306,7 @@ public class AbilitiesTab {
         // Contents
         JButton btnEnableAll = new CampaignOptionsButton("AddAll");
         btnEnableAll.addActionListener(e -> {
-            for (CampaignOptionsAbilityInfo abilityInfo : allAbilityInfo.values()) {
+            for (AbilityInfo abilityInfo : allAbilityInfo.values()) {
                 abilityInfo.setEnabled(true);
             }
 
@@ -317,7 +315,7 @@ public class AbilitiesTab {
 
         JButton btnRemoveAll = new CampaignOptionsButton("RemoveAll");
         btnRemoveAll.addActionListener(e -> {
-            for (CampaignOptionsAbilityInfo abilityInfo : allAbilityInfo.values()) {
+            for (AbilityInfo abilityInfo : allAbilityInfo.values()) {
                 abilityInfo.setEnabled(false);
             }
 
@@ -347,7 +345,7 @@ public class AbilitiesTab {
         Collections.sort(sortedAbilityNames);
 
         for (String abilityName : sortedAbilityNames) {
-            CampaignOptionsAbilityInfo abilityInfo = allAbilityInfo.get(abilityName);
+            AbilityInfo abilityInfo = allAbilityInfo.get(abilityName);
 
             if (abilityInfo.getCategory() == abilityCategory) {
                 JPanel abilityPanel = createSPAPanel(abilityInfo);
@@ -396,7 +394,7 @@ public class AbilitiesTab {
      *
      * @return A {@code JPanel} containing the UI elements related to the ability.
      */
-    private JPanel createSPAPanel(CampaignOptionsAbilityInfo abilityInfo) {
+    private JPanel createSPAPanel(AbilityInfo abilityInfo) {
         SpecialAbility ability = abilityInfo.getAbility();
 
         // Initialization
@@ -483,7 +481,7 @@ public class AbilitiesTab {
     private boolean editSPA(SpecialAbility ability) {
         Map<String, SpecialAbility> temporaryMap = new HashMap<>();
 
-        for (Entry<String, CampaignOptionsAbilityInfo> info : allAbilityInfo.entrySet()) {
+        for (Entry<String, AbilityInfo> info : allAbilityInfo.entrySet()) {
             temporaryMap.put(info.getKey(), info.getValue().getAbility());
         }
 
@@ -556,7 +554,7 @@ public class AbilitiesTab {
     public void applyCampaignOptionsToCampaign(@Nullable CampaignPreset preset) {
         Map<String, SpecialAbility> enabledAbilities = new HashMap<>();
 
-        for (CampaignOptionsAbilityInfo abilityInfo : allAbilityInfo.values()) {
+        for (AbilityInfo abilityInfo : allAbilityInfo.values()) {
             if (abilityInfo.isEnabled()) {
                 enabledAbilities.put(abilityInfo.getAbility().getName(), abilityInfo.getAbility());
             }
